+++ packages/analytics/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dexto/analytics
 
+## 1.8.5
+
+### Patch Changes
+
+- Updated dependencies [ee05731]
+    - @dexto/core@1.8.5
+    - @dexto/agent-management@1.8.5
+
 ## 1.8.4
 
 ### Patch Changes